:sparkles: feat: Added mappings to switch buffer by index and to cycle between them
Added `<Plug>`-prefixed configurable mappings to switch the focused
buffer by index and to go to the previous and next buffer in the
bufferline. Also added option to specify whether or not the cycling
should be strict, i.e. what happens when you're at the first (last) buffer
and you cycle to the previous (next) buffer.